Add files via upload
The committed files are created as a result of working on tasks of the first "ML OPS" by MADE homework. Some tasks have been completed, some stay unsolved.  The most complicated and interesting at the same time task in my own opinion is the "hydra" module. My experiments with it have taken quite a time but the result is a little bit under expectation (that's why I write that I deserve 1 ball for it).
All in all, it is a very interesting task.